Cellula_lastweek_Omar_Arnous/requirements.txt
langchain
langchain-community==0.3.7
langchain-core==0.3.17
langchain-openai==0.2.8
langgraph==0.2.33
duckduckgo-search
pydantic==2.9.2
python-dotenv==1.0.1
fastapi
uvicorn